the sum of numbers from 1 to N using getter and setter methods. - Introduce 'thirtySix' and 'thirtySeven' classes for animal sound demonstration. - Add 'thirtyTwo' class for circle and rectangle calculations. - Create 'one' class to demonstrate inheritance and method overriding in animal classes.

+++ assignment/ten.java
@@ -10,8 +10,8 @@ public class ten {
     public int multiply(int a, int b) {
         return a * b;
     }
-    public int divide(int a, int b) {
-        return a / b;
+    public double divide(int a, int b) {
+        return (double) a / b;
     }
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
